+++ src/App.css
@@ -126,167 +126,7 @@ button:hover {
 
 /* ============================================================== */
 
-.button-wrapper {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	gap: 1rem;
-}
-
-@media (min-width: 576px) {
-	.button-wrapper {
-		flex-direction: row;
-	}
-}
-
-.hero-row {
-	border-bottom-left-radius: 50px;
-	border-bottom-right-radius: 50px;
-	box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
-	background-color: #f6f8fb;
-	padding-top: 20px;
-	margin-bottom: 20px;
-}
-
-.hero-btn {
-	font-size: 1.2rem;
-	font-family: inherit;
-	border: 1px solid black;
-	cursor: pointer;
-	border-radius: 9px;
-	padding: 1rem 2rem;
-	text-decoration: none;
-	transition: background-color 500ms;
-}
-
-.hero-btn:hover {
-	background-color: #e0e4e9;
-	transform: translateY(-2px);
-	box-shadow: 0 4px 8px rgba(220, 215, 215, 0.2);
-}
-
-.button-subtitle {
-	font-size: 10px;
-}
-
-.button-title {
-	font-size: 26px;
-}
-
-.heading-primary {
-	font-size: 2.5rem;
-	color: #367ed0;
-	font-weight: 700;
-	letter-spacing: -0.5px;
-	line-height: 1.2;
-	margin-bottom: 1.6rem;
-	/* margin-top: 3rem; */
-
-	@media (min-width: 576px) {
-		font-size: 3.2rem;
-	}
-
-	@media (min-width: 768px) {
-		font-size: 3.8rem;
-	}
-
-	@media (min-width: 992px) {
-		font-size: 4.2rem;
-	}
-}
-
-.hero-description {
-	font-size: 1.2rem;
-	line-height: 1.5;
-	margin-bottom: 2.4rem;
-	@media (min-width: 576px) {
-		font-size: 1.4rem;
-	}
-
-	@media (min-width: 768px) {
-		font-size: 1.5rem;
-	}
-}
-
-.margin-right-btn {
-	margin-right: 1rem;
-
-	@media (min-width: 576px) {
-		margin-right: 1.5rem;
-	}
-
-	@media (min-width: 768px) {
-		margin-right: 2.4rem;
-	}
-}
-/* ================= */
-.hero-image-container {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-}
-
-.hero-img-box {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-}
-
-.hero-img {
-	max-width: 100%;
-	height: auto;
-}
-
-@media (max-width: 768px) {
-	.hero-img {
-		width: 80%;
-	}
-	.move-down-sm {
-		margin-top: 30px;
-	}
-}
-
-@media (max-width: 576px) {
-	.hero-img {
-		width: 60%;
-	}
-}
-
-@media (max-width: 576px) {
-	.hero-text-box {
-		padding-bottom: 30px !important;
-	}
-}
-
 /* ====================== cards =================================== */
-.custom-card {
-	border-radius: 10px;
-	box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
-	transition: box-shadow 0.3s ease-in-out;
-	width: 70%;
-	align-items: center;
-}
-
-.custom-card-img {
-	width: 60%;
-	border-top-left-radius: 10px;
-	border-top-right-radius: 10px;
-}
-
-.custom-card:hover {
-	box-shadow: 0px 15px 16px rgba(0, 0, 0, 0.2);
-}
-
-.custom-card-title {
-	font-size: 30px;
-	text-align: center !important;
-	font-weight: bolder;
-}
-
-.custom-card-text {
-	font-size: 20px;
-	text-align: center !important;
-}
 
 /*======================= how it works button ========================*/
 
@@ -313,7 +153,6 @@ button:hover {
 		border-radius: 40px !important;
 		box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
 		transition: box-shadow 0.3s ease-in-out;
-		width: 100% !important;
 	}
 
 	.hero-btn {
