+++ src/Components/Hero.js
@@ -1,0 +1,84 @@
+import React from "react";
+import { useTranslation } from "react-i18next";
+import { Container, Row, Col, Button } from "react-bootstrap";
+import "../index.css"; // Ensure to create and import your custom CSS for additional styling
+
+const Hero = () => {
+	const { t } = useTranslation();
+
+	return (
+		<Container fluid className="hero-section py-5">
+			<Row className="align-items-center">
+				<Col xl={8} lg={8} md={6} sm={12} xs={12} className="hero-text-container hero-text-box">
+					<h1 className="heading-primary">{t("Landing.landingh1")}</h1>
+					<p className="hero-description">{t("Landing.landingh2")}</p>
+
+					<div className="button-wrapper">
+						<Button
+							href="https://play.google.com/store/apps/details?id=com.afkanerd.sw0b"
+							target="_blank"
+							className="hero-btn btn btn--fill"
+							size="lg"
+							style={{
+								backgroundColor: "#367ed0",
+								color: "#FFFFFF",
+								justifyContent: "center",
+								width: "100%",
+								maxWidth: "200px"
+							}}
+						>
+							<div className="d-flex w-100">
+								<img
+									src="/playstore.svg"
+									alt="Android"
+									className="img-fluid logo"
+									style={{ maxWidth: "40px", marginRight: "10px" }}
+								/>
+								<div className="text-start">
+									<span className="button-subtitle fs-8">{t("Landing.AndroidSub")}</span>
+									<br />
+									<span className="button-title fs-4">{t("Landing.Android")}</span>
+								</div>
+							</div>
+						</Button>
+
+						<Button
+							href="https://play.google.com/store/apps/details?id=com.afkanerd.sw0b"
+							target="_blank"
+							className="hero-btn btn btn--outline"
+							size="lg"
+							style={{
+								backgroundColor: "white",
+								color: "#1c222c",
+								justifyContent: "center",
+								width: "100%",
+								maxWidth: "200px"
+							}}
+						>
+							<div className="d-flex w-100">
+								<img
+									src="/desktop.svg"
+									alt="Desktop"
+									className="img-fluid logo"
+									style={{ maxWidth: "40px", marginRight: "10px" }}
+								/>
+								<div className="text-start">
+									<span className="button-subtitle fs-8">{t("Landing.DesktopSub")}</span>
+									<br />
+									<span className="button-title fs-4">{t("Landing.Desktop")}</span>
+								</div>
+							</div>
+						</Button>
+					</div>
+				</Col>
+				<Col xl={4} lg={4} md={6} sm={12} xs={12} className="hero-image-container">
+					<div className="hero-img-box move-down-sm">
+						<img src="/phone3.png" alt="Relaysms App" className="hero-img img-fluid" />
+					</div>
+				</Col>
+			</Row>
+		</Container>
+	);
+};
+
+export default Hero;
