+++ src/index.css
@@ -1,0 +1,158 @@
+/*============ Howitworks.css============*/
+.howitworks-section {
+	background-color: #f8f9fa;
+	padding: 3rem 1rem;
+}
+
+.section-title {
+	font-size: calc(1rem + 1.5vw);
+	font-weight: bold;
+	margin-bottom: 1rem;
+	color: #343a40;
+}
+
+.section-subtitle {
+	font-size: calc(1.2rem + 0.5vw);
+	color: #6c757d;
+}
+
+.custom-card {
+	border: none;
+	overflow: hidden;
+	transition: transform 0.2s;
+	border-radius: 20px !important;
+	box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
+	transition: box-shadow 0.3s ease-in-out;
+	width: 60%;
+	align-items: center;
+}
+
+.custom-card:hover {
+	transform: scale(1.05);
+}
+
+.getting-started-link {
+	font-size: 1.5rem;
+	color: #024995;
+	font-weight: 600;
+
+	text-decoration: none;
+}
+
+.getting-started-link:hover {
+	text-decoration: none;
+	color: #e79405;
+	font-weight: 800;
+	font-size: x-large;
+}
+
+.arrow-icon {
+	margin-left: 10px;
+	transform: rotate(-30deg);
+}
+
+.custom-card-img {
+	width: 60%;
+	border-top-left-radius: 10px;
+	border-top-right-radius: 10px;
+}
+
+.custom-card-title {
+	font-size: 30px;
+	text-align: center !important;
+	font-weight: bolder;
+}
+
+.custom-card-text {
+	font-size: 20px;
+	text-align: center !important;
+	color: #6c757d;
+}
+/* ================ hero section ========================== */
+.hero-section {
+	background-color: #f8f9fa;
+	padding: 3rem 1rem;
+}
+
+.hero-text-box {
+	padding: 2rem;
+}
+
+.heading-primary {
+	font-size: 2.5rem;
+	color: #367ed0;
+	font-weight: 700;
+	letter-spacing: -0.5px;
+	line-height: 1.2;
+	margin-bottom: 1.6rem;
+
+	@media (min-width: 576px) {
+		font-size: 3.2rem;
+	}
+
+	@media (min-width: 768px) {
+		font-size: 3.8rem;
+	}
+
+	@media (min-width: 992px) {
+		font-size: 4.2rem;
+	}
+}
+
+.hero-description {
+	font-size: 1.2rem;
+	line-height: 1.5;
+	margin-bottom: 2.4rem;
+	color: #6c757d;
+	@media (min-width: 576px) {
+		font-size: 1.4rem;
+	}
+
+	@media (min-width: 768px) {
+		font-size: 1.5rem;
+	}
+}
+
+.button-wrapper {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	gap: 1rem;
+}
+
+@media (min-width: 576px) {
+	.button-wrapper {
+		flex-direction: row;
+	}
+}
+
+.button-subtitle {
+	font-size: 10px;
+}
+
+.button-title {
+	font-size: 26px;
+}
+
+.margin-right-btn {
+	margin-right: 1rem;
+
+	@media (min-width: 576px) {
+		margin-right: 1.5rem;
+	}
+
+	@media (min-width: 768px) {
+		margin-right: 2.4rem;
+	}
+}
+.hero-img-box {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
+
+.hero-img {
+	max-width: 100%;
+	height: auto;
+}
+/* -------------------------end of hero section------------------------- */
